shadow_consumer/requirements.txt
kafka-python-ng
requests
elasticsearch==8.5.3